+++ src/Hooks/useDocumentTitle.jsx
@@ -1,3 +1,4 @@
+import { useEffect } from "react";
 export const useDocumentTitle = (docTitle) => {
     useEffect(() => {
         document.title = docTitle;
